# api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Kurator API",
    "version": "1.0",
    "description": "REST API for Kurator: collections, items, search, external metadata lookup, session-based auth, bundled SQL migrations with first-boot bootstrap, and optional S3-backed image uploads.\n\n**API v2:** **`/api/v2/hitlists`** adds permalink slugs, internet-**public** visibility, stub entries, votes, and comments (see tag hitlists-v2).\n\n**Authentication:** Sessions are opaque tokens stored hashed in the database. The API accepts (1) the **kurator_session** HTTP-only cookie (typical browser flow via the Next.js same-origin proxy) or (2) **`Authorization: Bearer <session_token>`** on any protected route. When both are present, the **cookie wins**. After **POST /api/v1/auth/register**, **POST /api/v1/auth/login** (when 2FA is not required), or **POST /api/v1/auth/login/2fa**, the JSON body includes **session_token** \u2014 the same secret as the cookie \u2014 for native and non-cookie clients. **POST /api/v1/auth/logout** revokes the session for either mechanism; browsers also receive `Set-Cookie` clearing the cookie."
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
...
  "securityDefinitions": {
    "SessionCookie": {
      "type": "apiKey",
      "name": "kurator_session",
      "in": "cookie",
      "description": "HTTP-only session cookie set on successful register/login/2fa. Preferred when both cookie and Authorization Bearer are present."
    },
    "BearerToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer followed by the opaque session_token from POST /auth/login, /auth/login/2fa, or /auth/register (same secret as the kurator_session cookie). Example: Bearer <session_token>"
    }
  },
  "paths": {
...
    "/api/v1/me/custom-theme": {
      "get": {
        "tags": ["custom-themes"],
        "summary": "Get my custom theme",
        "description": "Requires Kurator Pro (**SessionCookie** or **BearerToken**). Returns saved YAML or defaults.",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "produces": ["application/json"],
        "responses": {
          "200": {"description": "OK"},
          "403": {"description": "Pro required"},
          "401": {"description": "Unauthorized"}
        }
      },
      "put": {
        "tags": ["custom-themes"],
        "summary": "Save custom theme YAML",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{"name": "body", "in": "body", "required": true, "schema": {"$ref": "#/definitions/CustomThemeYAMLBody"}}],
        "responses": {
          "200": {"description": "OK"},
          "400": {"description": "Validation errors"},
          "403": {"description": "Pro required"},
          "429": {"description": "Rate limited (10 uploads/day)"}
        }
      },
      "delete": {
        "tags": ["custom-themes"],
        "summary": "Reset custom theme",
        "description": "Removes saved draft YAML. Fails with 409 if the theme is still published (unpublish first).",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "produces": ["application/json"],
        "responses": {
          "200": {"description": "OK"},
          "403": {"description": "Pro required"},
          "409": {"description": "Theme still published"}
        }
      }
    },
    "/api/v1/me/custom-theme/unpublish": {
      "post": {
        "tags": ["custom-themes"],
        "summary": "Unpublish custom theme",
        "description": "Removes all published marketplace versions. Clears active theme for the owner and affected Pro users; notifies users who were actively using the theme.",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "produces": ["application/json"],
        "responses": {
          "200": {"description": "Unpublished"},
          "400": {"description": "Theme not published"},
          "403": {"description": "Pro required"}
        }
      }
    },
    "/api/v1/me/custom-theme/created": {
      "delete": {
        "tags": ["custom-themes"],
        "summary": "Delete created custom theme",
        "description": "Deletes the user's theme draft and own library entry. Requires the theme to be unpublished first.",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "produces": ["application/json"],
        "responses": {
          "200": {"description": "Deleted"},
          "403": {"description": "Pro required"},
          "409": {"description": "Theme still published"}
        }
      }
    },
    "/api/v1/me/custom-theme/validate": {
      "post": {
        "tags": ["custom-themes"],
        "summary": "Validate custom theme YAML",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{"name": "body", "in": "body", "required": true, "schema": {"$ref": "#/definitions/CustomThemeYAMLBody"}}],
        "responses": {
          "200": {"description": "Validation result"},
          "403": {"description": "Pro required"}
        }
      }
    },
    "/api/v1/me/custom-theme/publish": {
      "post": {
        "tags": ["custom-themes"],
        "summary": "Publish custom theme",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "produces": ["application/json"],
        "responses": {
          "200": {"description": "Published theme"},
          "403": {"description": "Pro required"}
        }
      }
    },
    "/api/v1/me/custom-theme/google-fonts": {
      "get": {
        "tags": ["custom-themes"],
        "summary": "List Google Font families for theme editor autocomplete",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "produces": ["application/json"],
        "responses": {
          "200": {"description": "Font family names"},
          "403": {"description": "Pro required"}
        }
      }
    },
    "/api/v1/me/custom-theme/library": {
      "get": {
        "tags": ["custom-themes"],
        "summary": "List custom theme library",
        "description": "Requires Kurator Pro. Returns own saved theme and marketplace installs.",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "produces": ["application/json"],
        "responses": {
          "200": {"description": "Library entries"},
          "403": {"description": "Pro required"}
        }
      },
      "post": {
        "tags": ["custom-themes"],
        "summary": "Install marketplace theme",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{"name": "body", "in": "body", "required": true, "schema": {"type": "object", "properties": {"published_theme_id": {"type": "string"}}}}],
        "responses": {
          "200": {"description": "Installed library entry"},
          "403": {"description": "Pro required"},
          "409": {"description": "Already installed"}
        }
      }
    },
    "/api/v1/me/custom-theme/library/{id}": {
      "delete": {
        "tags": ["custom-themes"],
        "summary": "Remove marketplace theme from library",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "produces": ["application/json"],
        "parameters": [{"name": "id", "in": "path", "required": true, "type": "string"}],
        "responses": {
          "200": {"description": "Removed"},
          "400": {"description": "Cannot remove own theme entry"},
          "403": {"description": "Pro required"},
          "404": {"description": "Library entry not found"}
        }
      }
    },
    "/api/v1/me/custom-theme/active": {
      "get": {
        "tags": ["custom-themes"],
        "summary": "Get active custom theme YAML",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "produces": ["application/json"],
        "responses": {
          "200": {"description": "Active theme or inactive"},
          "403": {"description": "Pro required"}
        }
      },
      "patch": {
        "tags": ["custom-themes"],
        "summary": "Set active custom theme",
        "description": "Pass library_id as UUID string or null to clear.",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{"name": "body", "in": "body", "required": true, "schema": {"type": "object", "properties": {"library_id": {"type": "string", "x-nullable": true}}}}],
        "responses": {
          "200": {"description": "Active theme updated"},
          "403": {"description": "Pro required"},
          "404": {"description": "Library entry not found"}
        }
      }
    },
    "/api/v1/custom-themes": {
      "get": {
        "tags": ["custom-themes"],
        "summary": "Browse published custom themes",
        "produces": ["application/json"],
        "parameters": [
          {"name": "q", "in": "query", "type": "string"},
          {"name": "limit", "in": "query", "type": "integer"},
          {"name": "offset", "in": "query", "type": "integer"}
        ],
        "responses": {"200": {"description": "Theme list"}}
      }
    },
    "/api/v1/custom-themes/{id}": {
      "get": {
        "tags": ["custom-themes"],
        "summary": "Get published custom theme",
        "produces": ["application/json"],
        "parameters": [{"name": "id", "in": "path", "required": true, "type": "string"}],
        "responses": {"200": {"description": "OK"}, "404": {"description": "Not found"}}
      }
    },
    "/api/v1/custom-themes/{id}/report": {
      "post": {
        "tags": ["custom-themes"],
        "summary": "Report published custom theme",
        "security": [{"SessionCookie": []}, {"BearerToken": []}],
        "consumes": ["application/json"],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"},
          {"name": "body", "in": "body", "required": true, "schema": {"$ref": "#/definitions/CustomThemeReportBody"}}
        ],
        "responses": {"200": {"description": "Report queued"}}
      }
    },
...
  },
  "definitions": {
...
    "CustomThemeYAMLBody": {
      "type": "object",
      "required": ["yaml"],
      "properties": {
        "yaml": {
          "type": "string",
          "description": "customTheme v1 YAML document"
        }
      }
    },
    "CustomThemeReportBody": {
      "type": "object",
      "required": ["reason"],
      "properties": {
        "reason": {
          "type": "string",
          "maxLength": 500
        }
      }
    },
...
  }
}